fix all the warnings

+++ Core/include/Acts/Utilities/detail/Axis.hpp
@@ -287,7 +287,8 @@ class Axis<AxisType::Equidistant, bdt> final : public IAxis {
   /// @note Bin indices start at @c 1. The underflow bin has the index @c 0
   ///       while the index <tt>nBins + 1</tt> indicates the overflow bin .
   size_t getBin(ActsScalar x) const {
-    return wrapBin(std::floor((x - getMin()) / getBinWidth()) + 1);
+    return wrapBin(
+        static_cast<int>(std::floor((x - getMin()) / getBinWidth()) + 1));
   }
 
   /// @brief get bin width
