+++ Examples/Detectors/ContextualDetector/include/ActsExamples/ContextualDetector/AlignedDetector.hpp
@@ -30,7 +30,7 @@ class AlignedDetector : public ActsExamples::IBaseDetector {
 
   struct Config : public GenericDetector::Config {
     /// Seed for the decorator random numbers.
-    double seed = 1324354657;
+    size_t seed = 1324354657;
     /// Size of a valid IOV.
     size_t iovSize = 100;
     /// Span until garbage collection is active.
